Updated docs, refactored the UMI sharing part.
